+++ NEWS.md
@@ -14,6 +14,13 @@
 - *Internal* I added `try_safely()` to minimize parsing errors from breaking
   the complete pdf processing pipeline. I also muffled parsing of pdf text 
   tables with `quietly()`.
+  
+- The *county deaths* label was changed to *deaths* in the afternoon of 
+  between 2020-03-21 11:26:37 EDT and 2020-03-21 16:26:06 EDT.
+  
+- *Internal* Use the negatives reported in the PDF and the positive, pending,
+  and deaths reported in the dashboard for the plots on the front page. Also
+  replace change in "resolved" cases with plot of new deaths.
 
 # 2020-03-21